backend/requirements.txt
torch==2.3.1
transformers==4.40.2
accelerate==0.30.1
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
python-dotenv==1.0.1
slowapi==0.1.9
structlog==24.1.0
prometheus-fastapi-instrumentator==6.1.0
gunicorn==21.2.0
sseclient-py==1.8.0
